+++ frontend/svelte/env.config.mjs
@@ -67,6 +67,8 @@ const OWN_CANISTER_ID = getRequiredEnvVar("OWN_CANISTER_ID");
 const GOVERNANCE_CANISTER_URL = getRequiredEnvVar("GOVERNANCE_CANISTER_URL");
 const LEDGER_CANISTER_URL = getRequiredEnvVar("LEDGER_CANISTER_URL");
 const OWN_CANISTER_URL = getRequiredEnvVar("OWN_CANISTER_URL");
+const WASM_CANISTER_ID = getRequiredEnvVar("WASM_CANISTER_ID");
+console.log({ WASM_CANISTER_ID, path: "env.config.mjs" });
 
 // Configuration
 // ... The testnet name
@@ -92,6 +94,7 @@ export const envConfig = {
   LEDGER_CANISTER_URL,
   ROLLUP_WATCH,
   ENABLE_NEW_SPAWN_FEATURE,
+  WASM_CANISTER_ID,
 };
 
 // Note: This is a useful printout at build time.  Please do not remove.