+++ frontend/svelte/src/lib/components/launchpad/Proposals.svelte
@@ -1,22 +1,23 @@
 <script lang="ts">
-  import type { ProposalInfo } from "@dfinity/nns";
   import { onMount } from "svelte";
   import { listSnsProposals } from "../../services/sns.services";
   import { i18n } from "../../stores/i18n";
+  import {
+    openForVotesSnsProposalsStore,
+    snsProposalsStore,
+  } from "../../stores/projects.store";
+  import { isNullable } from "../../utils/utils";
   import CardGrid from "../ui/CardGrid.svelte";
   import SkeletonProposalCard from "../ui/SkeletonProposalCard.svelte";
   import ProposalCard from "./ProposalCard.svelte";
 
   let loading: boolean = false;
-  let proposals: ProposalInfo[] | undefined = undefined;
+  $: loading = isNullable($snsProposalsStore);
 
-  const load = async () => {
-    loading = true;
-
-    // TODO L2-751: replace the source
-    proposals = await listSnsProposals();
-
-    loading = false;
+  const load = () => {
+    if ($snsProposalsStore === undefined) {
+      listSnsProposals();
+    }
   };
 
   onMount(load);
@@ -27,16 +28,14 @@
     <SkeletonProposalCard />
     <SkeletonProposalCard />
   </CardGrid>
-{:else if proposals !== undefined}
-  {#if proposals.length === 0}
-    <p class="no-proposals">{$i18n.voting.nothing_found}</p>
-  {:else}
-    <CardGrid>
-      {#each proposals as proposalInfo (proposalInfo.id)}
-        <ProposalCard {proposalInfo} />
-      {/each}
-    </CardGrid>
-  {/if}
+{:else if $openForVotesSnsProposalsStore.length === 0}
+  <p class="no-proposals">{$i18n.voting.nothing_found}</p>
+{:else}
+  <CardGrid>
+    {#each $openForVotesSnsProposalsStore as proposalInfo (proposalInfo.id)}
+      <ProposalCard {proposalInfo} />
+    {/each}
+  </CardGrid>
 {/if}
 
 <style lang="scss">