+++ frontend/svelte/src/lib/components/neuron-detail/actions/SpawnNeuronButton.svelte
@@ -1,13 +1,17 @@
 <script lang="ts">
-  import { ICP, type NeuronInfo } from "@dfinity/nns";
-  import { MIN_NEURON_STAKE } from "../../../constants/neurons.constants";
+  import type { NeuronInfo } from "@dfinity/nns";
+  import { E8S_PER_ICP } from "../../../constants/icp.constants";
+  import {
+    MIN_NEURON_STAKE,
+    SPAWN_VARIANCE_PERCENTAGE,
+  } from "../../../constants/neurons.constants";
   import SpawnNeuronModal from "../../../modals/neurons/SpawnNeuronModal.svelte";
   import { accountsStore } from "../../../stores/accounts.store";
   import { i18n } from "../../../stores/i18n";
+  import { formatNumber, formatPercentage } from "../../../utils/format.utils";
   import { replacePlaceholders } from "../../../utils/i18n.utils";
-  import { formatICP } from "../../../utils/icp.utils";
   import {
-    isEnoughToStakeNeuron,
+    isEnoughMaturityToSpawn,
     isNeuronControlledByHardwareWallet,
   } from "../../../utils/neuron.utils";
   import Tooltip from "../../ui/Tooltip.svelte";
@@ -27,8 +31,9 @@
   $: enoughMaturity =
     neuron.fullNeuron === undefined
       ? false
-      : isEnoughToStakeNeuron({
-          stake: ICP.fromE8s(neuron.fullNeuron.maturityE8sEquivalent),
+      : isEnoughMaturityToSpawn({
+          neuron,
+          percentage: 100,
         });
 </script>
 
@@ -42,7 +47,18 @@
     text={replacePlaceholders(
       $i18n.neuron_detail.spawn_maturity_disabled_tooltip,
       {
-        $amount: formatICP({ value: BigInt(MIN_NEURON_STAKE), detailed: true }),
+        $amount: formatNumber(
+          MIN_NEURON_STAKE / E8S_PER_ICP / SPAWN_VARIANCE_PERCENTAGE,
+          { minFraction: 4, maxFraction: 4 }
+        ),
+        $min: formatNumber(MIN_NEURON_STAKE / E8S_PER_ICP, {
+          minFraction: 0,
+          maxFraction: 0,
+        }),
+        $varibility: formatPercentage(SPAWN_VARIANCE_PERCENTAGE, {
+          minFraction: 0,
+          maxFraction: 0,
+        }),
       }
     )}
   >