+++ frontend/svelte/src/lib/stores/projects.store.ts
@@ -1,7 +1,10 @@
+import type { ProposalInfo } from "@dfinity/nns";
 import type { Principal } from "@dfinity/principal";
 import { SnsSwapLifecycle } from "@dfinity/sns";
 import { derived, writable, type Readable } from "svelte/store";
 import type { SnsSummary, SnsSwapCommitment } from "../types/sns";
+import { isProposalOpenForVotes } from "../utils/proposals.utils";
+import { isNullable } from "../utils/utils";
 
 export type SnsSummariesStore =
   | {
@@ -26,6 +29,50 @@ export interface SnsFullProject {
 }
 
 export type SnsFullProjectsStore = SnsFullProject[] | undefined;
+
+export type SnsProposalsStore =
+  | {
+      proposals: ProposalInfo[];
+      certified: boolean;
+    }
+  | undefined
+  | null;
+
+const initSnsProposalsStore = () => {
+  const { subscribe, set } = writable<SnsProposalsStore>(undefined);
+
+  return {
+    subscribe,
+
+    reset() {
+      set(undefined);
+    },
+
+    setLoadingState() {
+      set(null);
+    },
+
+    setProposals({
+      proposals,
+      certified,
+    }: {
+      proposals: ProposalInfo[];
+      certified: boolean;
+    }) {
+      set({
+        proposals,
+        certified,
+      });
+    },
+  };
+};
+
+const initOpenForVotesSnsProposalsStore = () =>
+  derived([snsProposalsStore], ([$snsProposalsStore]): ProposalInfo[] =>
+    isNullable($snsProposalsStore)
+      ? []
+      : $snsProposalsStore.proposals.filter(isProposalOpenForVotes)
+  );
 
 const initSnsSummariesStore = () => {
   const { subscribe, set } = writable<SnsSummariesStore>(undefined);
@@ -92,6 +139,10 @@ const initSnsSwapCommitmentsStore = () => {
 // used to improve loading state display only
 export const snsesCountStore = writable<number | undefined>(undefined);
 
+export const snsProposalsStore = initSnsProposalsStore();
+export const openForVotesSnsProposalsStore =
+  initOpenForVotesSnsProposalsStore();
+
 export const snsSummariesStore = initSnsSummariesStore();
 export const snsSwapCommitmentsStore = initSnsSwapCommitmentsStore();
 
@@ -146,10 +197,25 @@ const filterProjectsStore = ({
 export const openProjectsStore = derived(
   snsFullProjectsStore,
   ($snsFullProjectsStore: SnsFullProject[] | undefined) =>
-    filterProjectsStore({
-      swapLifecycle: SnsSwapLifecycle.Open,
-      $snsFullProjectsStore,
-    })
+    // TODO: redo after demo
+    // filterProjectsStore({
+    //   swapLifecycle: SnsSwapLifecycle.Open,
+    //   $snsFullProjectsStore,
+    // })
+    $snsFullProjectsStore === undefined
+      ? undefined
+      : $snsFullProjectsStore.filter(
+          ({
+            summary: {
+              swap: {
+                state: { lifecycle, open_time_window },
+              },
+            },
+          }) =>
+            lifecycle === SnsSwapLifecycle.Open ||
+            (lifecycle === SnsSwapLifecycle.Pending &&
+              open_time_window.length > 0)
+        )
 );
 
 export const committedProjectsStore = derived(