+++ frontend/svelte/src/lib/types/i18n.d.ts
@@ -34,6 +34,7 @@ interface I18nError {
   proposal_payload: string;
   proposal_payload_not_found: string;
   neuron_not_found: string;
+  neuron_spawning: string;
   neuron_load: string;
   list_proposals: string;
   list_canisters: string;
@@ -210,6 +211,7 @@ interface I18nNeurons {
   create: string;
   edit_source: string;
   status_locked: string;
+  status_spawning: string;
   status_unspecified: string;
   status_dissolved: string;
   status_dissolving: string;
@@ -235,6 +237,7 @@ interface I18nNeurons {
   edit_delay: string;
   merge_neurons_article_title: string;
   cannot_merge_neuron_community: string;
+  cannot_merge_neuron_spawning: string;
   cannot_merge_neuron_hotkey: string;
   cannot_merge_hardware_wallet: string;
   only_merge_two: string;
@@ -279,6 +282,8 @@ interface I18nFollow_neurons {
   topic_9_subtitle: string;
   topic_10_title: string;
   topic_10_subtitle: string;
+  topic_11_title: string;
+  topic_11_subtitle: string;
   current_followees: string;
   add: string;
 }
@@ -360,6 +365,7 @@ interface I18nTopics {
   NetworkCanisterManagement: string;
   Kyc: string;
   NodeProviderRewards: string;
+  SnsDecentralizationSale: string;
 }
 
 interface I18nRewards {
@@ -539,12 +545,14 @@ interface I18nNeuron_detail {
   spawn_maturity_confirmation_q: string;
   spawn_maturity_confirmation_a: string;
   spawn_maturity_choose: string;
-  spawn_maturity_explanation: string;
-  spawn_maturity_explanation_hw: string;
+  spawn_maturity_explanation_1: string;
+  spawn_maturity_explanation_2: string;
+  spawn_maturity_note_hw: string;
   current_stake: string;
   current_maturity: string;
   dissolve_delay_range: string;
   maturity_range: string;
+  spawning_neuron_info: string;
   maturity_percentage: string;
 }