+++ frontend/svelte/src/tests/lib/components/launchpad/Proposals.spec.ts
@@ -6,9 +6,9 @@ import type { ProposalInfo } from "@dfinity/nns";
 import { render, waitFor } from "@testing-library/svelte";
 import Proposals from "../../../../lib/components/launchpad/Proposals.svelte";
 import { listSnsProposals } from "../../../../lib/services/sns.services";
+import { snsProposalsStore } from "../../../../lib/stores/projects.store";
 import en from "../../../mocks/i18n.mock";
 import { mockProposalInfo } from "../../../mocks/proposal.mock";
-import { mockWaiting } from "../../../mocks/utils.mock";
 
 jest.mock("../../../../lib/services/sns.services", () => {
   return {
@@ -17,10 +17,12 @@ jest.mock("../../../../lib/services/sns.services", () => {
 });
 
 describe("Proposals", () => {
-  const mockProposals = (proposals: ProposalInfo[]) =>
-    (listSnsProposals as jest.Mock).mockImplementation(() =>
-      mockWaiting(0.1, proposals)
-    );
+  const mockProposals = (proposals: ProposalInfo[] | null) =>
+    proposals === null
+      ? snsProposalsStore.setLoadingState()
+      : snsProposalsStore.setProposals({ proposals, certified: true });
+
+  beforeEach(snsProposalsStore.reset);
 
   afterAll(jest.clearAllMocks);
 
@@ -30,8 +32,19 @@ describe("Proposals", () => {
     expect(listSnsProposals).toBeCalled();
   });
 
+  it("should not trigger listSnsProposals if already loaded", () => {
+    snsProposalsStore.setProposals({
+      proposals: [],
+      certified: true,
+    });
+
+    render(Proposals);
+
+    expect(listSnsProposals).toBeCalled();
+  });
+
   it("should display skeletons", async () => {
-    mockProposals([]);
+    mockProposals(null);
 
     const { container } = render(Proposals);